Added Exceptions and Valdiations

+++ Assignment2/src/main/java/com/neu/edu/controller/BillController.java
@@ -4,6 +4,7 @@ import java.text.SimpleDateFormat;
 import java.util.Base64;
 import java.util.Date;
 import java.util.List;
+import java.util.UUID;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -12,25 +13,34 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.HttpClientErrorException.BadRequest;
 
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import com.google.gson.JsonObject;
+import com.neu.edu.exception.BillException;
 import com.neu.edu.model.Bill;
 import com.neu.edu.model.PaymentStatus;
 import com.neu.edu.model.User;
 import com.neu.edu.repository.BillRepository;
 import com.neu.edu.repository.BillRepositoryfindaSpecificBill;
 import com.neu.edu.repository.UserRepository;
+
 
 @RestController
 public class BillController {
@@ -50,11 +60,15 @@ public class BillController {
 	String pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
 	SimpleDateFormat simpleDateFormat = new SimpleDateFormat(pattern);
 
+	
+    
 	//Post a bill by authenticating User
 	//201 created, 400 bad request, 401 for no auth or unauthorized
 	@PostMapping(value= "/v1/bill")
-	public ResponseEntity<?> postBillByUserId(@Valid @RequestBody Bill bill, HttpServletRequest request, HttpServletResponse response)
+	public ResponseEntity<?> postBillByUserId(@Validated @RequestBody(required = false) Bill bill, HttpServletRequest request, HttpServletResponse response)
 	{
+		
+		
 		String authorization = request.getHeader("Authorization");
 		JsonObject entity = new JsonObject();
 		if(authorization != null && authorization.toLowerCase().startsWith("basic"))
@@ -72,56 +86,70 @@ public class BillController {
 			User user = userRepository.findByemail(email);
 			if(user == null)
 			{
-				entity.addProperty("message", "Please enter correct Username or Password");
+				entity.addProperty("message", "User does not exist. Please enter correct Username or Password");
 			}
 			else if(user != null && !bCryptPasswordEncoder.matches(password, user.getPassword()))
 			{
 				entity.addProperty("message", "The Password is Invalid");
-				return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+				return new ResponseEntity<String>(entity.toString() , HttpStatus.UNAUTHORIZED);
 			}
 			else
-			{		
-				if( (bill.getVendor()!=null && bill.getVendor().trim().length() >0) && bill.getBilldate()!=null  && bill.getDuedate()!=null && 
-					(bill.getAmountdue()>0.00 && bill.getAmountdue()< Double.MAX_VALUE)
-					&& bill.getCategories().size()>0 && bill.getCategories()!= null && bill.getPaymentStatus() != null )
+			{	
+				if(bill==null)
 				{
-					if(validateDate(bill.getBilldate()) && validateDate(bill.getDuedate()))
+					entity.addProperty("message", "The Request Body cannot be null");
+					return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+				}
+				if( (bill.getVendor()!=null && bill.getVendor().trim().length() >0) && bill.getBilldate()!=null  && bill.getDuedate()!=null && 
+					(bill.getAmountdue()>=0.01 && bill.getAmountdue()< Double.MAX_VALUE)
+					&&  bill.getCategories()!= null && bill.getPaymentStatus() != null )
+				{
+					if(bill.getCategories().size()>0 )
 					{
-						if(bill.getPaymentStatus().equals(PaymentStatus.paid) || bill.getPaymentStatus().equals(PaymentStatus.due) ||  bill.getPaymentStatus().equals(PaymentStatus.no_payment_required) || bill.getPaymentStatus().equals(PaymentStatus.past_due))
+						if(validateDate(bill.getBilldate()) && validateDate(bill.getDuedate()))
 						{
-							String dateFormat = simpleDateFormat.format(new Date());	
-							Bill b = new Bill();
-							b.setUser(user);
-							b.setVendor(bill.getVendor());
-							b.setBilldate(bill.getBilldate());
-							b.setDuedate(bill.getDuedate());
-							b.setAmountdue(bill.getAmountdue());
-							b.setCategories(bill.getCategories());
-							b.setCreated_ts(dateFormat.toString());
-							b.setUpdated_ts(dateFormat.toString());
-							b.setPaymentStatus(bill.getPaymentStatus());
-							b.setOwner_id(user.getId());
-							billRepository.save(b);
-							b.setUser(null);
-							return new ResponseEntity<Bill>(b , HttpStatus.OK);
+							if(bill.getPaymentStatus().equals(PaymentStatus.paid) || bill.getPaymentStatus().equals(PaymentStatus.due) ||  bill.getPaymentStatus().equals(PaymentStatus.no_payment_required) || bill.getPaymentStatus().equals(PaymentStatus.past_due))
+							{
+								String dateFormat = simpleDateFormat.format(new Date());	
+								Bill b = new Bill();
+								b.setUser(user);
+								b.setVendor(bill.getVendor());
+								b.setBilldate(bill.getBilldate());
+								b.setDuedate(bill.getDuedate());
+								b.setAmountdue(bill.getAmountdue());
+								b.setCategories(bill.getCategories());
+								b.setCreated_ts(dateFormat.toString());
+								b.setUpdated_ts(dateFormat.toString());
+								b.setPaymentStatus(bill.getPaymentStatus());
+								b.setOwner_id(user.getId());
+								billRepository.save(b);
+								b.setUser(null);
+								return new ResponseEntity<Bill>(b , HttpStatus.CREATED);
+							}
+							else
+							{
+								entity.addProperty("validation", " Please enter a valid Payment Status 	");
+								return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+							}
+							
 						}
 						else
 						{
-							entity.addProperty("validation", " Please enter a valid Payment Status 	");
+							entity.addProperty("validation", " Please enter a valid date in YYYY-MM-DD format ");
 							return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
 						}
-						
 					}
 					else
 					{
-						entity.addProperty("validation", " Please enter a valid date in YYYY-MM-DD format ");
+						entity.addProperty("validation", " Categories cannot be empty ");
 						return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
 					}
+				
 					
 				}
 				else
 				{
-					entity.addProperty("message", "vendor, bill_date, due_date, amount_due, categories or payment status cannot be empty and amount_due should be greater than 0.00");
+					entity.addProperty("message", "vendor, bill_date, due_date, amount_due, categories, payment status cannot be empty or Amount cannot be below 0.00");
 					return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
 				}
 			
@@ -160,12 +188,12 @@ public class BillController {
 			User user = userRepository.findByemail(email);
 			if(user == null)
 			{
-				entity.addProperty("message", "Please enter correct Username or Password");
+				entity.addProperty("message", "User does not exist. Please enter correct Username or Password");
 			}
 			else if(user != null && !bCryptPasswordEncoder.matches(password, user.getPassword()))
 			{
 				entity.addProperty("message", "The Password is Invalid");
-				return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+				return new ResponseEntity<String>(entity.toString() , HttpStatus.UNAUTHORIZED);
 			}
 			else
 			{
@@ -189,8 +217,9 @@ public class BillController {
 
 	//Get Specific bill by billId
 	@GetMapping(value = "/v1/bill/{id}")
-	public ResponseEntity<?> getSingleBillbyId(HttpServletRequest request, HttpServletResponse response, @PathVariable(value="id") @NotBlank @NotNull String billId)
+	public ResponseEntity<?> getSingleBillbyId(HttpServletRequest request, HttpServletResponse response, @PathVariable(value="id" ) String billId)
 	{
+		
 		String authorization = request.getHeader("Authorization");
 		JsonObject entity = new JsonObject();
 		if(authorization != null && authorization.toLowerCase().startsWith("basic"))
@@ -209,16 +238,31 @@ public class BillController {
 			User user = userRepository.findByemail(email);
 			if(user == null)
 			{
-				entity.addProperty("message", "Please enter correct Username or Password");
+				entity.addProperty("message", "User does not exist. Please enter correct Username or Password");
 			}
 			else if(user != null && !bCryptPasswordEncoder.matches(password, user.getPassword()))
 			{
 				entity.addProperty("message", "The Password is Invalid");
-				return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+				return new ResponseEntity<String>(entity.toString() , HttpStatus.UNAUTHORIZED);
 			}
 			else
 			{
+				 UUID  uid  = null ;
+		            try 
+		            {
+		                 uid = UUID.fromString(billId);
+		                 System.out.println("Bill UUID is : ");
+		            }
+		            catch (Exception e)
+		            {
+
+						entity.addProperty("message", "The bill does not exist.");
+						return new ResponseEntity<String>(entity.toString(), HttpStatus.NOT_FOUND);
+		            }
+		            
 				List<Bill> listOfBills = billRepository.findByUserId(user.getId());
+				 
+				
 				Bill bill = billRepositoryfindaSpecificBill.findById(billId);
 				if(listOfBills.size() > 0)
 				{
@@ -245,7 +289,6 @@ public class BillController {
 				{
 					entity.addProperty("message", "The bill does not exist.");
 					return new ResponseEntity<String>(entity.toString(), HttpStatus.NOT_FOUND);
-					
 				}
 				
 				
@@ -262,7 +305,7 @@ public class BillController {
 	
 	//Update a bill 
 	@PutMapping(value = "/v1/bill/{id}")
-	public ResponseEntity<?> updateBillById(HttpServletRequest request, HttpServletResponse response, @RequestBody Bill bill, @PathVariable(value = "id") @NotBlank @NotNull String billId )
+	public ResponseEntity<?> updateBillById(HttpServletRequest request, HttpServletResponse response, @RequestBody(required = false) Bill bill, @PathVariable(required = true, value = "id") @NotBlank @NotNull String billId )
 	{
 		String authorization = request.getHeader("Authorization");
 		JsonObject entity = new JsonObject();
@@ -282,17 +325,36 @@ public class BillController {
 			User user = userRepository.findByemail(email);
 			if(user == null)
 			{
-				entity.addProperty("message", "Please enter correct Username or Password");
+				entity.addProperty("message", "User does not exist. Please enter correct Username or Password");
 			}
 			else if(user != null && !bCryptPasswordEncoder.matches(password, user.getPassword()))
 			{
 				entity.addProperty("message", "The Password is Invalid");
-				return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+				return new ResponseEntity<String>(entity.toString() ,HttpStatus.UNAUTHORIZED);
 			}
 			else
 			{
-				List<Bill> listOfBills = billRepository.findByUserId(user.getId());
+				if(bill==null)
+				{
+					entity.addProperty("message", "The Request Body cannot be null");
+					return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+				}
 				
+				List<Bill> listOfBills = billRepository.findByUserId(user.getId());
+				 UUID  uid  = null ;
+		            try 
+		            {
+		                 uid = UUID.fromString(billId);
+		                 System.out.println("Bill UUID is : ");
+		            }
+		            catch (Exception e)
+		            {
+
+						entity.addProperty("message", "The bill does not exist or bill Id is incorrect.");
+						return new ResponseEntity<String>(entity.toString(), HttpStatus.NOT_FOUND);
+		            }
+		            
+			//&& 
 				if(billId != null )
 				{
 					Bill b = billRepositoryfindaSpecificBill.findById(billId);
@@ -303,8 +365,10 @@ public class BillController {
 							if(listOfBills.contains(b))
 							{
 								if( (bill.getVendor()!=null && bill.getVendor().trim().length() >0) && bill.getBilldate()!=null  && bill.getDuedate()!=null && 
-										(bill.getAmountdue()>0.00 && bill.getAmountdue()< Double.MAX_VALUE)
-										&& bill.getCategories().size()>0 && bill.getCategories()!= null  && bill.getPaymentStatus() != null )
+										(bill.getAmountdue()>=0.01 && bill.getAmountdue()< Double.MAX_VALUE)
+										&& bill.getCategories() != null  && bill.getPaymentStatus() != null )
+									{
+									if(bill.getCategories().size()>0 )
 									{
 										if(validateDate(bill.getBilldate()) && validateDate(bill.getDuedate()))
 										{
@@ -329,6 +393,13 @@ public class BillController {
 											return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
 										}
 									}
+									else
+									{
+										entity.addProperty("validation", " categories cannot be empty ");
+										return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+									}
+										
+									}
 								else
 								{
 									entity.addProperty("message", "vendor, bill_date, due_date, amount_due, categories or payment status cannot be empty");
@@ -345,7 +416,7 @@ public class BillController {
 						else
 						{
 							entity.addProperty("message", "The bill does not exist.");
-							return new ResponseEntity<String>(entity.toString(), HttpStatus.BAD_REQUEST);
+							return new ResponseEntity<String>(entity.toString(), HttpStatus.NOT_FOUND);
 							
 						}
 						
@@ -394,21 +465,40 @@ public class BillController {
 			User user = userRepository.findByemail(email);
 			if(user == null)
 			{
-				entity.addProperty("message", "Please enter correct Username or Password");
+				entity.addProperty("message", "User does not exist. Please enter correct Username or Password");
 			}
 			else if(user != null && !bCryptPasswordEncoder.matches(password, user.getPassword()))
 			{
 				entity.addProperty("message", "The Password is Invalid");
-				return new ResponseEntity<String>(entity.toString() , HttpStatus.BAD_REQUEST);
+				return new ResponseEntity<String>(entity.toString() , HttpStatus.UNAUTHORIZED);
 			}
 			else
 			{
 				List<Bill> listOfBills = billRepository.findByUserId(user.getId());
+				 UUID  uid  = null ;
+		            try 
+		            {
+		                 uid = UUID.fromString(billId);
+		                 System.out.println("Bill UUID is : ");
+		            }
+		            catch (Exception e)
+		            {
+
+						entity.addProperty("message", "The bill does not exist or bill Id is incorrect.");
+						return new ResponseEntity<String>(entity.toString(), HttpStatus.NOT_FOUND);
+		            }
+		            
 				if(billId != null )
 				{
 					if(listOfBills.size() > 0)
 					{	
 						Bill b = billRepositoryfindaSpecificBill.findById(billId);
+						if(b==null)
+						{
+							entity.addProperty("message", "The bill does not exist.");
+							return new ResponseEntity<String>(entity.toString(), HttpStatus.NOT_FOUND);
+							
+						}
 						if(listOfBills.contains(b))
 						{			
 							billRepository.delete(b);
@@ -443,15 +533,17 @@ public class BillController {
 		entity.addProperty("message", "Invalid. Unable to Authenticate");	
 		return new ResponseEntity<String>(entity.toString() , HttpStatus.UNAUTHORIZED);
 	}
-	
-	
+
+
+	  
+    
 	public Boolean validateDate(String date) 
 	{
 		if (date != null || (!date.equalsIgnoreCase(""))) 
 		{
-			String datevalidator = "^[0-9]{4}([- \\/.])(((0[13578]|(10|12))\\1(0[1-9]|[1-2][0-9]|3[0-1]))|(02\\1(0[1-9]|[1-2][0-9]))|((0[469]|11)\\1(0[1-9]|[1-2][0-9]|30)))$";
+			String datevalidator = "^[12]\\d{3}([- \\/.])(((0[13578]|(10|12))\\1(0[1-9]|[1-2][0-9]|3[0-1]))|(02\\1(0[1-9]|[1-2][0-9]))|((0[469]|11)\\1(0[1-9]|[1-2][0-9]|30)))$";
 			//^\d{4}\-(0[1-9]|1[012])\-(0[1-9]|[12][0-9]|3[01])$
-			
+			//[12]\\d{3}
 
 			return date.matches(datevalidator);
 		}
